# assets/openapi.json
{
    "openapi": "3.0.0",
    "servers": [
        {
            "url": "http://localhost:3001/api/",
            "description": "Official Solarcord Instance"
        }
    ],
    "info": {
        "description": "Solarcord is a free open source selfhostable discord compatible chat, voice and video platform",
        "version": "1.0.0",
        "title": "Solarcord HTTP API Routes",
        "termsOfService": "",
        "contact": {
            "name": "Solarcord"
        },
        "license": {
            "name": "AGPLV3",
            "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "description": "Bearer/Bot prefixes are not required."
            }
        },
...
    },
...
    "paths": {
...
        "/invites/{code}": {
            "get": {
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "code"
                    }
                ],
                "tags": [
                    "invites"
                ]
            },
            "post": {
                "x-right-required": "USE_MASS_INVITES",
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "code"
                    }
                ],
                "tags": [
                    "invites"
                ]
            },
            "delete": {
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "code"
                    }
                ],
                "tags": [
                    "invites"
                ]
            }
        },
...
    }
}